Fix compiler error on VS2015 in recent commit

+++ modules/juce_audio_processors/format_types/juce_VST3PluginFormat.cpp
@@ -2046,7 +2046,7 @@ public:
         JUCE_DECLARE_VST3_COM_QUERY_METHODS
 
         Steinberg::int32 PLUGIN_API getParameterCount() override                                { return numQueuesUsed; }
-        Vst::IParamValueQueue* PLUGIN_API getParameterData (Steinberg::int32 index) override    { return isPositiveAndBelow (index, numQueuesUsed) ? queues[(int) index] : nullptr; }
+        Vst::IParamValueQueue* PLUGIN_API getParameterData (Steinberg::int32 index) override    { return isPositiveAndBelow (static_cast<int> (index), numQueuesUsed) ? queues[(int) index] : nullptr; }
 
         Vst::IParamValueQueue* PLUGIN_API addParameterData (const Vst::ParamID& id, Steinberg::int32& index) override
         {
